azuracast_migration/backend/requirements.txt
django>=5.0
django-extensions
whitenoise
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-decouple
requests
drf-spectacular
celery
redis
mutagen
python-magic
ffmpeg-python
dj-database-url
django-guardian
django-auditlog
Pillow
pyotp
qrcode
pyloudnorm
soundfile
channels[daphne]
channels_redis
django-storages
boto3
django-prometheus
docker
pytz
asgiref
